+++ src/main/java/org/flickit/flickitassessmentcore/application/service/answer/GetAnswerListService.java
@@ -3,7 +3,7 @@ package org.flickit.flickitassessmentcore.application.service.answer;
 import lombok.RequiredArgsConstructor;
 import org.flickit.flickitassessmentcore.application.domain.crud.PaginatedResponse;
 import org.flickit.flickitassessmentcore.application.port.in.answer.GetAnswerListUseCase;
-import org.flickit.flickitassessmentcore.application.port.out.LoadAnswersByQuestionnaireIdPort;
+import org.flickit.flickitassessmentcore.application.port.out.answer.LoadAnswersByQuestionnaireIdPort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
